Extract COBOL reference helper
